feat: add resolve command
The ESM `import.meta.url` value does not work reliably in environments
like electron just yet so add a resolve command to let us work out
paths to files in dependencies.
